Activate zoom when cursor is moving over the image whilst page is being reloaded
If the cursor is moving over the image area whilst the page is still being loaded, zoom will not trigger. You will need to move the cursor of the image and then move your cursor into the image in order to activate the zoom. My tiny addition fixes this bug.